backend/app/services/token_service.py
# services/token_service.py

from datetime import datetime, timedelta
from jose import jwt
from app.core.config import SECRET_KEY, ALGORITHM, ACCESS_TOKEN_EXPIRE_MINUTES

def create_access_token(data: dict, role : str):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({
        "exp": expire,
        "role": role  #  rolü ekliyoruz
        })
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
