backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]  # JWT işlemleri için
passlib[bcrypt]            # Şifre hashing
sqlalchemy                 # Veritabanı işlemleri
pydantic[email]                  # Veri doğrulama için (FastAPI zaten kullanıyor)